+++ ra-back/src/main/java/pl/raptors/ra_back/domain/mapsMetadata/MapRotation.java
@@ -1,0 +1,8 @@
+package pl.raptors.ra_back.domain.mapsMetadata;
+
+import lombok.Data;
+
+@Data
+public class MapRotation {
+    double yaw;
+}