+++ ra-back/src/main/java/pl/raptors/ra_back/domain/mapsMetadata/MapTransformation.java
@@ -1,0 +1,10 @@
+package pl.raptors.ra_back.domain.mapsMetadata;
+
+import lombok.Data;
+
+@Data
+public class MapTransformation {
+    double x;
+    double y;
+    double yaw;
+}