+++ ra-front/src/app/components/map/corridors/corridor-list/corridor-list.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CorridorListComponent } from './corridor-list.component';
 
-describe('CorridorListComponent', () => {
+xdescribe('CorridorListComponent', () => {
   let component: CorridorListComponent;
   let fixture: ComponentFixture<CorridorListComponent>;