+++ ra-front/src/app/model/GenericRobotModel/SpecialProperty/SpecialPropertyEnum.ts
@@ -1,0 +1,7 @@
+export enum SpecialPropertyEnum {
+  AVAILABLE = 'available',
+  POSITION = 'position',
+  ORIENTATION = 'orientation',
+  BATTERY_LEVEL = 'batteryLevel',
+  IP = 'robotIp'
+}